frontend for now

+++ shop-backend/src/main/java/com/github/reck0nerrr/shop/controllers/UserController.java
@@ -1,12 +1,16 @@
 package com.github.reck0nerrr.shop.controllers;
 
 import com.github.reck0nerrr.shop.dtos.UserDtos.*;
+import com.github.reck0nerrr.shop.security.SecurityUtils;
+import com.github.reck0nerrr.shop.security.UserPrincipal;
 import com.github.reck0nerrr.shop.service.UserService;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
@@ -28,12 +32,14 @@ public class UserController {
     }
 
     @GetMapping
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<List<UserResponse>> getAll() {
         return ResponseEntity.ok(userService.getAll());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id,@AuthenticationPrincipal UserPrincipal principal) {
+        SecurityUtils.requireSelfOrAdmin(principal, id);
         userService.delete(id);
         return ResponseEntity.noContent().build();
     }
